+++ src/app/components/core/components.module.ts
