+++ src/app/components/core/footer/footer.component.html
@@ -3,7 +3,7 @@
     <ion-grid class="landingPage">
       <ion-row class="ion-align-items-center">
         <ion-col size="12" size-md="1">
-          <img src="../../../assets/prototypefund.svg" alt="Prototype Fund Logo" id="prototypefundLogo" />
+          <img src="../../../../assets/prototypefund.svg" alt="Prototype Fund Logo" id="prototypefundLogo" />
         </ion-col>
         <ion-col size="12" size-md="5">
           <ion-text>
