+++ src/app/components/core/footer/footer.component.spec.ts
