+++ src/app/components/core/footer/footer.component.ts
