+++ src/app/components/core/header/header.component.scss
