+++ src/app/pages/home/home.module.ts
@@ -7,7 +7,7 @@ import {ReactiveFormsModule} from '@angular/forms';
 import {HomePage} from './home.page';
 
 import {HomePageRoutingModule} from './home-routing.module';
-import {ComponentsModule} from 'src/app/components/components.module';
+import {ComponentsModule} from 'src/app/components/core/components.module';
 
 @NgModule({
   imports: [CommonModule, FormsModule, IonicModule, HomePageRoutingModule, ComponentsModule, ReactiveFormsModule],
