+++ src/app/pages/infosite/infosite.module.ts
@@ -8,7 +8,7 @@ import {InfositePageRoutingModule} from './infosite-routing.module';
 
 import {InfositePage} from './infosite.page';
 
-import {ComponentsModule} from 'src/app/components/components.module';
+import {ComponentsModule} from 'src/app/components/core/components.module';
 
 @NgModule({
   imports: [CommonModule, FormsModule, IonicModule, InfositePageRoutingModule, ComponentsModule],
