+++ src/app/pages/wizard/start/start.module.ts
@@ -7,7 +7,7 @@ import {IonicModule} from '@ionic/angular';
 import {StartPageRoutingModule} from './start-routing.module';
 
 import {StartPage} from './start.page';
-import {ComponentsModule} from '../../../components/components.module';
+import {ComponentsModule} from '../../../components/core/components.module';
 
 @NgModule({
   imports: [CommonModule, FormsModule, IonicModule, StartPageRoutingModule, ComponentsModule],
